Reduce code duplication in existing XHFP
