+++ src/sodiumCompatibility/java/net/coderbot/iris/compat/sodium/impl/vertex_format/QuadView.java
@@ -1,7 +1,104 @@
 package net.coderbot.iris.compat.sodium.impl.vertex_format;
 
-public interface QuadView {
-	float x(int index);
-	float y(int index);
-	float z(int index);
+import net.coderbot.iris.vertices.NormalHelper;
+
+public abstract class QuadView {
+	public abstract float x(int index);
+	public abstract float y(int index);
+	public abstract float z(int index);
+	public abstract float u(int index);
+	public abstract float v(int index);
+
+	public int computeTangent(float normalX, float normalY, float normalZ) {
+		// Capture all of the relevant vertex positions
+		float x0 = x(0);
+		float y0 = y(0);
+		float z0 = z(0);
+
+		float x1 = x(1);
+		float y1 = y(1);
+		float z1 = z(1);
+
+		float x2 = x(2);
+		float y2 = y(2);
+		float z2 = z(2);
+
+		float edge1x = x1 - x0;
+		float edge1y = y1 - y0;
+		float edge1z = z1 - z0;
+
+		float edge2x = x2 - x0;
+		float edge2y = y2 - y0;
+		float edge2z = z2 - z0;
+
+		float u0 = u(0);
+		float v0 = v(0);
+
+		float u1 = u(1);
+		float v1 = v(1);
+
+		float u2 = u(2);
+		float v2 = v(2);
+
+		float deltaU1 = u1 - u0;
+		float deltaV1 = v1 - v0;
+		float deltaU2 = u2 - u0;
+		float deltaV2 = v2 - v0;
+
+		float fdenom = deltaU1 * deltaV2 - deltaU2 * deltaV1;
+		float f;
+
+		if (fdenom == 0.0) {
+			f = 1.0f;
+		} else {
+			f = 1.0f / fdenom;
+		}
+
+		float tangentx = f * (deltaV2 * edge1x - deltaV1 * edge2x);
+		float tangenty = f * (deltaV2 * edge1y - deltaV1 * edge2y);
+		float tangentz = f * (deltaV2 * edge1z - deltaV1 * edge2z);
+		float tcoeff = rsqrt(tangentx * tangentx + tangenty * tangenty + tangentz * tangentz);
+		tangentx *= tcoeff;
+		tangenty *= tcoeff;
+		tangentz *= tcoeff;
+
+		float bitangentx = f * (-deltaU2 * edge1x + deltaU1 * edge2x);
+		float bitangenty = f * (-deltaU2 * edge1y + deltaU1 * edge2y);
+		float bitangentz = f * (-deltaU2 * edge1z + deltaU1 * edge2z);
+		float bitcoeff = rsqrt(bitangentx * bitangentx + bitangenty * bitangenty + bitangentz * bitangentz);
+		bitangentx *= bitcoeff;
+		bitangenty *= bitcoeff;
+		bitangentz *= bitcoeff;
+
+		// predicted bitangent = tangent × normal
+		// Compute the determinant of the following matrix to get the cross product
+		//  i  j  k
+		// tx ty tz
+		// nx ny nz
+
+		float pbitangentx = tangenty * normalZ - tangentz * normalY;
+		float pbitangenty = -(tangentx * normalZ - tangentz * normalX);
+		float pbitangentz = tangentx * normalX - tangenty * normalY;
+
+		float dot = (bitangentx * pbitangentx) + (bitangenty * pbitangenty) + (bitangentz * pbitangentz);
+		float tangentW;
+
+		if (dot < 0) {
+			tangentW = -1.0F;
+		} else {
+			tangentW = 1.0F;
+		}
+
+		return NormalHelper.packNormal(tangentx, tangenty, tangentz, tangentW);
+	}
+
+	private static float rsqrt(float value) {
+		if (value == 0.0f) {
+			// You heard it here first, folks: 1 divided by 0 equals 1
+			// In actuality, this is a workaround for normalizing a zero length vector (leaving it as zero length)
+			return 1.0f;
+		} else {
+			return (float) (1.0 / Math.sqrt(value));
+		}
+	}
 }
