+++ src/sodiumCompatibility/java/net/coderbot/iris/compat/sodium/impl/vertex_format/terrain_xhfp/QuadViewTerrain.java
@@ -4,7 +4,7 @@ import net.coderbot.iris.compat.sodium.impl.vertex_format.QuadView;
 
 import java.nio.ByteBuffer;
 
-public class QuadViewTerrain implements QuadView {
+public class QuadViewTerrain extends QuadView {
 	public ByteBuffer buffer;
 	public int writeOffset;
 	private static final int STRIDE = 36;
@@ -19,6 +19,20 @@ public class QuadViewTerrain implements QuadView {
 
 	public float z(int index) {
 		return normalizeVertexPositionShortAsFloat(buffer.getShort(writeOffset + 4 - STRIDE * (3 - index)));
+	}
+
+	@Override
+	public float u(int index) {
+		return normalizeVertexTextureShortAsFloat(buffer.getShort(writeOffset + 12 - STRIDE * (3 - index)));
+	}
+
+	@Override
+	public float v(int index) {
+		return normalizeVertexTextureShortAsFloat(buffer.getShort(writeOffset + 14 - STRIDE * (3 - index)));
+	}
+
+	private static float normalizeVertexTextureShortAsFloat(short value) {
+		return (value & 0xFFFF) * (1.0f / 32768.0f);
 	}
 
 	// TODO: Verify that this works with the new changes to the CVF
