+++ src/sodiumCompatibility/java/net/coderbot/iris/compat/sodium/impl/vertex_format/terrain_xhfp/XHFPModelVertexBufferWriterNio.java
@@ -13,6 +13,7 @@ import net.coderbot.iris.vendored.joml.Vector3f;
 
 import java.nio.ByteBuffer;
 
+// TODO: Add Unsafe version
 public class XHFPModelVertexBufferWriterNio extends VertexBufferWriterNio implements ModelVertexSink, MaterialIdAwareVertexWriter {
     private MaterialIdHolder idHolder;
 
@@ -111,87 +112,7 @@ public class XHFPModelVertexBufferWriterNio extends VertexBufferWriterNio implem
             buffer.putInt(i + 28 - STRIDE * 2, packedNormal);
             buffer.putInt(i + 28 - STRIDE * 3, packedNormal);
 
-            // Capture all of the relevant vertex positions
-            float x0 = normalizeVertexPositionShortAsFloat(buffer.getShort(i - STRIDE * 3));
-            float y0 = normalizeVertexPositionShortAsFloat(buffer.getShort(i + 2 - STRIDE * 3));
-            float z0 = normalizeVertexPositionShortAsFloat(buffer.getShort(i + 4 - STRIDE * 3));
-
-            float x1 = normalizeVertexPositionShortAsFloat(buffer.getShort(i - STRIDE * 2));
-            float y1 = normalizeVertexPositionShortAsFloat(buffer.getShort(i + 2 - STRIDE * 2));
-            float z1 = normalizeVertexPositionShortAsFloat(buffer.getShort(i + 4 - STRIDE * 2));
-
-            float x2 = normalizeVertexPositionShortAsFloat(buffer.getShort(i - STRIDE));
-            float y2 = normalizeVertexPositionShortAsFloat(buffer.getShort(i + 2 - STRIDE));
-            float z2 = normalizeVertexPositionShortAsFloat(buffer.getShort(i + 4 - STRIDE));
-
-            float edge1x = x1 - x0;
-            float edge1y = y1 - y0;
-            float edge1z = z1 - z0;
-
-            float edge2x = x2 - x0;
-            float edge2y = y2 - y0;
-            float edge2z = z2 - z0;
-
-            float u0 = normalizeVertexTextureShortAsFloat(buffer.getShort(i + 12 - STRIDE * 3));
-            float v0 = normalizeVertexTextureShortAsFloat(buffer.getShort(i + 14 - STRIDE * 3));
-
-            float u1 = normalizeVertexTextureShortAsFloat(buffer.getShort(i + 12 - STRIDE * 2));
-            float v1 = normalizeVertexTextureShortAsFloat(buffer.getShort(i + 14 - STRIDE * 2));
-
-            float u2 = normalizeVertexTextureShortAsFloat(buffer.getShort(i + 12 - STRIDE));
-            float v2 = normalizeVertexTextureShortAsFloat(buffer.getShort(i + 14 - STRIDE));
-
-            float deltaU1 = u1 - u0;
-            float deltaV1 = v1 - v0;
-            float deltaU2 = u2 - u0;
-            float deltaV2 = v2 - v0;
-
-            float fdenom = deltaU1 * deltaV2 - deltaU2 * deltaV1;
-            float f;
-
-            if (fdenom == 0.0) {
-                f = 1.0f;
-            } else {
-                f = 1.0f / fdenom;
-            }
-
-            float tangentx = f * (deltaV2 * edge1x - deltaV1 * edge2x);
-            float tangenty = f * (deltaV2 * edge1y - deltaV1 * edge2y);
-            float tangentz = f * (deltaV2 * edge1z - deltaV1 * edge2z);
-            float tcoeff = rsqrt(tangentx * tangentx + tangenty * tangenty + tangentz * tangentz);
-            tangentx *= tcoeff;
-            tangenty *= tcoeff;
-            tangentz *= tcoeff;
-
-            float bitangentx = f * (-deltaU2 * edge1x + deltaU1 * edge2x);
-            float bitangenty = f * (-deltaU2 * edge1y + deltaU1 * edge2y);
-            float bitangentz = f * (-deltaU2 * edge1z + deltaU1 * edge2z);
-            float bitcoeff = rsqrt(bitangentx * bitangentx + bitangenty * bitangenty + bitangentz * bitangentz);
-            bitangentx *= bitcoeff;
-            bitangenty *= bitcoeff;
-            bitangentz *= bitcoeff;
-
-            // predicted bitangent = tangent × normal
-            // Compute the determinant of the following matrix to get the cross product
-            //  i  j  k
-            // tx ty tz
-            // nx ny nz
-
-            float pbitangentx =   tangenty * normal.z() - tangentz * normal.y();
-            float pbitangenty = -(tangentx * normal.z() - tangentz * normal.x());
-            float pbitangentz =   tangentx * normal.x() - tangenty * normal.y();
-
-            float dot = (bitangentx * pbitangentx) + (bitangenty * pbitangenty) + (bitangentz * pbitangentz);
-            byte tangentW;
-
-            if (dot < 0) {
-                tangentW = -127;
-            } else {
-                tangentW = 127;
-            }
-
-            int tangent = Norm3b.pack(tangentx, tangenty, tangentz);
-            tangent |= (tangentW << 24);
+            int tangent = currentQuad.computeTangent(normal.x(), normal.y(), normal.z());
 
             buffer.putInt(i + 24, tangent);
             buffer.putInt(i + 24 - STRIDE, tangent);
@@ -200,26 +121,6 @@ public class XHFPModelVertexBufferWriterNio extends VertexBufferWriterNio implem
         }
 
         this.advance();
-    }
-
-    // TODO: Verify that this works with the new changes to the CVF
-    private static float normalizeVertexPositionShortAsFloat(short value) {
-        return (value & 0xFFFF) * (1.0f / 65535.0f);
-    }
-
-    // TODO: Verify that this is correct
-    private static float normalizeVertexTextureShortAsFloat(short value) {
-        return (value & 0xFFFF) * (1.0f / 32768.0f);
-    }
-
-    private static float rsqrt(float value) {
-        if (value == 0.0f) {
-            // You heard it here first, folks: 1 divided by 0 equals 1
-            // In actuality, this is a workaround for normalizing a zero length vector (leaving it as zero length)
-            return 1.0f;
-        } else {
-            return (float) (1.0 / Math.sqrt(value));
-        }
     }
 
     @Override
